Add dynamic metafile path
In some situations it may be useful to have multiple sets of metadata
for the same hierarchy (e.g. representing how a file set should appear on
different hosts).
The ability to select a metadata file may be useful in this case.

Patch by Sergio Callegari <sergio.callegari@gmail.com>