metaentry.c: Define PATH_MAX if it's undefined.
PATH_MAX is not guaranteed by POSIX to be present and some systems, like
GNU Hurd, indeed does not define it.

http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/limits.h.html
https://www.gnu.org/software/hurd/hurd/porting/guidelines.html

For now let's define PATH_MAX as 4096 (if it is not already defined),
because that's how Linux defines it and it seems quite reasonable.
In future dependence on PATH_MAX can be removed.