README.md: Create a symlink to README.
README is actually written in markdown, but I don't like .md extension.
README.md is here only for better experience of web viewers of the repo,
it's set to be not exported by git archive and won't be present in
released tarballs.

BTW if GitHub is offline, you can use following unofficial mirror:
http://repo.or.cz/metastore.git
(It also renders README.md file, that's why I mentioned it here.)